+++ docs/src/examples/components/Alert/Rtl/AlertExampleDismissAction.rtl.tsx
@@ -1,0 +1,20 @@
+import * as React from 'react'
+import { Alert } from '@stardust-ui/react'
+
+const AlertExampleDismissActionRtl = () => (
+  <Alert
+    actions={[
+      {
+        content: 'مرحبا',
+        primary: true,
+      },
+      {
+        content: 'عالم',
+      },
+    ]}
+    icon="exclamation-triangle"
+    content="مرحبا العالم"
+  />
+)
+
+export default AlertExampleDismissActionRtl