+++ packages/react/src/components/Menu/MenuItem.tsx
@@ -1,3 +1,5 @@
+import { Accessibility, menuItemBehavior, submenuBehavior } from '@stardust-ui/accessibility'
+import { focusAsync } from '@stardust-ui/react-bindings'
 import { EventListener } from '@stardust-ui/react-component-event-listener'
 import { Ref, toRefObject } from '@stardust-ui/react-component-ref'
 import * as customPropTypes from '@stardust-ui/react-proptypes'
@@ -22,8 +24,6 @@ import {
 import Icon, { IconProps } from '../Icon/Icon'
 import Menu, { MenuProps, MenuShorthandKinds } from './Menu'
 import Box, { BoxProps } from '../Box/Box'
-import { menuItemBehavior, submenuBehavior } from '../../lib/accessibility'
-import { Accessibility } from '../../lib/accessibility/types'
 import {
   ComponentEventHandler,
   WithAsProp,
@@ -31,7 +31,6 @@ import {
   ShorthandCollection,
   withSafeTypeForAs,
 } from '../../types'
-import { focusAsync } from '../../lib/accessibility/FocusZone'
 import { Popper } from '../../lib/positioner'
 
 export interface MenuItemSlotClassNames {