+++ packages/react/src/components/Toolbar/ToolbarRadioGroup.tsx
@@ -18,8 +18,11 @@ import {
 import { mergeComponentVariables } from '../../lib/mergeThemes'
 
 import { ShorthandCollection, WithAsProp, withSafeTypeForAs } from '../../types'
-import { Accessibility } from '../../lib/accessibility/types'
-import { toolbarRadioGroupBehavior, toolbarRadioGroupItemBehavior } from '../../lib/accessibility'
+import {
+  Accessibility,
+  toolbarRadioGroupBehavior,
+  toolbarRadioGroupItemBehavior,
+} from '@stardust-ui/accessibility'
 
 import ToolbarDivider from './ToolbarDivider'
 import ToolbarItem, { ToolbarItemProps } from './ToolbarItem'