+++ packages/react/src/lib/renderComponent.tsx
@@ -1,10 +1,21 @@
+import {
+  AccessibilityDefinition,
+  FocusZoneMode,
+  FocusZoneDefinition,
+  Accessibility,
+} from '@stardust-ui/accessibility'
+import {
+  callable,
+  FocusZone,
+  FocusZoneProps,
+  FOCUSZONE_WRAP_ATTRIBUTE,
+  getElementType,
+  getUnhandledProps,
+} from '@stardust-ui/react-bindings'
 import cx from 'classnames'
 import * as React from 'react'
 import * as _ from 'lodash'
 
-import callable from './callable'
-import getElementType from './getElementType'
-import getUnhandledProps from './getUnhandledProps'
 import logProviderMissingWarning from './providerMissingHandler'
 import {
   ComponentStyleFunctionParam,
@@ -14,22 +25,15 @@ import {
   PropsWithVarsAndStyles,
   State,
   ThemePrepared,
-  ComponentSlotStylesInput,
 } from '../themes/types'
 import { Props, ProviderContextPrepared } from '../types'
-import {
-  AccessibilityDefinition,
-  FocusZoneMode,
-  FocusZoneDefinition,
-  Accessibility,
-} from './accessibility/types'
 import { ReactAccessibilityBehavior, AccessibilityActionHandlers } from './accessibility/reactTypes'
 import getKeyDownHandlers from './getKeyDownHandlers'
 import { emptyTheme, mergeComponentStyles, mergeComponentVariables } from './mergeThemes'
-import { FocusZoneProps, FocusZone } from './accessibility/FocusZone'
-import { FOCUSZONE_WRAP_ATTRIBUTE } from './accessibility/FocusZone/focusUtilities'
 import createAnimationStyles from './createAnimationStyles'
-import Debug, { isEnabled as isDebugEnabled } from './debug'
+import { isEnabled as isDebugEnabled } from './debug/debugEnabled'
+import { DebugData } from './debug/debugData'
+import withDebugId from './withDebugId'
 
 export interface RenderResultConfig<P> {
   ElementType: React.ElementType<P>
@@ -46,14 +50,13 @@ export type RenderComponentCallback<P> = (config: RenderResultConfig<P>) => any
 
 export interface RenderConfig<P> {
   className?: string
-  defaultProps?: { [key: string]: any }
   displayName: string
   handledProps: string[]
   props: PropsWithVarsAndStyles
   state: State
   actionHandlers: AccessibilityActionHandlers
   render: RenderComponentCallback<P>
-  saveDebug: (debug: Debug | null) => void
+  saveDebug: (debug: DebugData | null) => void
 }
 
 const emptyBehavior: ReactAccessibilityBehavior = {
@@ -144,23 +147,12 @@ const renderWithFocusZone = <P extends {}>(
   return render(config)
 }
 
-const resolveStyles = (
-  styles: ComponentSlotStylesInput,
-  styleParam: ComponentStyleFunctionParam,
-): ComponentSlotStylesPrepared => {
-  return Object.keys(styles).reduce(
-    (acc, next) => ({ ...acc, [next]: callable(styles[next])(styleParam) }),
-    {},
-  )
-}
-
 const renderComponent = <P extends {}>(
   config: RenderConfig<P>,
   context?: ProviderContextPrepared,
 ): React.ReactElement<P> => {
   const {
     className,
-    defaultProps,
     displayName,
     handledProps,
     props,
@@ -177,13 +169,13 @@ const renderComponent = <P extends {}>(
   const { disableAnimations = false, renderer = null, rtl = false, theme = emptyTheme } =
     context || {}
 
-  const ElementType = getElementType({ defaultProps }, props) as React.ReactType<P>
+  const ElementType = getElementType(props) as React.ReactType<P>
   const stateAndProps = { ...state, ...props }
 
   // Resolve variables for this component, allow props.variables to override
   const resolvedVariables: ComponentVariablesObject = mergeComponentVariables(
     theme.componentVariables[displayName],
-    props.variables,
+    props.variables && withDebugId(props.variables, 'props.variables'),
   )(theme.siteVariables)
 
   const animationCSSProp = props.animation
@@ -193,9 +185,9 @@ const renderComponent = <P extends {}>(
   // Resolve styles using resolved variables, merge results, allow props.styles to override
   const mergedStyles: ComponentSlotStylesPrepared = mergeComponentStyles(
     theme.componentStyles[displayName],
-    { root: props.design },
-    { root: props.styles },
-    { root: animationCSSProp },
+    withDebugId({ root: props.design }, 'props.design'),
+    withDebugId({ root: props.styles }, 'props.styles'),
+    withDebugId({ root: animationCSSProp }, 'props.animation'),
   )
 
   const accessibility: ReactAccessibilityBehavior = getAccessibility(
@@ -205,7 +197,7 @@ const renderComponent = <P extends {}>(
     rtl,
   )
 
-  const unhandledProps = getUnhandledProps({ handledProps }, props)
+  const unhandledProps = getUnhandledProps(handledProps, props)
 
   const styleParam: ComponentStyleFunctionParam = {
     displayName,
@@ -222,13 +214,20 @@ const renderComponent = <P extends {}>(
   const direction = rtl ? 'rtl' : 'ltr'
   const felaParam = {
     theme: { direction },
+    displayName, // does not affect styles, only used by useEnhancedRenderer in docs
   }
 
   const resolvedStyles: ComponentSlotStylesPrepared = {}
+  const resolvedStylesDebug: { [key: string]: { styles: Object }[] } = {}
   const classes: ComponentSlotClasses = {}
 
   Object.keys(mergedStyles).forEach(slotName => {
     resolvedStyles[slotName] = callable(mergedStyles[slotName])(styleParam)
+
+    if (process.env.NODE_ENV !== 'production' && isDebugEnabled) {
+      resolvedStylesDebug[slotName] = resolvedStyles[slotName]['_debug']
+      delete resolvedStyles[slotName]['_debug']
+    }
 
     if (renderer) {
       classes[slotName] = renderer.renderRule(callable(resolvedStyles[slotName]), felaParam)
@@ -248,22 +247,40 @@ const renderComponent = <P extends {}>(
     theme,
   }
 
-  if (accessibility.focusZone) {
-    return renderWithFocusZone(render, accessibility.focusZone, resolvedConfig)
+  // conditionally add sources for evaluating debug information to component
+  if (process.env.NODE_ENV !== 'production' && isDebugEnabled) {
+    saveDebug({
+      componentName: displayName,
+      componentVariables: _.filter(
+        resolvedVariables._debug,
+        variables => !_.isEmpty(variables.resolved),
+      ),
+      componentStyles: _.mapValues(resolvedStylesDebug, v =>
+        _.filter(v, v => {
+          return !_.isEmpty(v.styles)
+        }),
+      ),
+      siteVariables: _.filter(theme.siteVariables._debug, siteVars => {
+        if (_.isEmpty(siteVars) || _.isEmpty(siteVars.resolved)) {
+          return false
+        }
+
+        const keys = Object.keys(siteVars.resolved)
+        if (
+          keys.length === 1 &&
+          keys.pop() === 'fontSizes' &&
+          _.isEmpty(siteVars.resolved['fontSizes'])
+        ) {
+          return false
+        }
+
+        return true
+      }),
+    })
   }
 
-  // conditionally add sources for evaluating debug information to component
-  if (isDebugEnabled) {
-    saveDebug(
-      new Debug({
-        componentName: displayName,
-        themes: context ? context.originalThemes : [],
-        instanceStylesOverrides: props.styles,
-        instanceVariablesOverrides: props.variables,
-        resolveStyles: styles => resolveStyles(styles, styleParam),
-        resolveVariables: variables => callable(variables)(theme.siteVariables),
-      }),
-    )
+  if (accessibility.focusZone) {
+    return renderWithFocusZone(render, accessibility.focusZone, resolvedConfig)
   }
 
   return render(resolvedConfig)