+++ packages/react/test/specs/components/Chat/Chat-test.ts
@@ -1,10 +1,9 @@
+import { chatBehavior, AccessibilityDefinition } from '@stardust-ui/accessibility'
 import { handlesAccessibility, isConformant } from 'test/specs/commonTests'
 
 import Chat from 'src/components/Chat/Chat'
 import implementsCollectionShorthandProp from '../../commonTests/implementsCollectionShorthandProp'
 import ChatItem from 'src/components/Chat/ChatItem'
-import { chatBehavior } from 'src/lib/accessibility'
-import { AccessibilityDefinition } from 'src/lib/accessibility/types'
 
 const chatImplementsCollectionShorthandProp = implementsCollectionShorthandProp(Chat)